Вернул класс LevelConverter в FactMap
