--- Depmon.Server.Collector/ILevelConverter.cs
@@ -1,8 +1,0 @@
-﻿using CsvHelper.TypeConversion;
-
-namespace Depmon.Server.Collector
-{
-    public interface ILevelConverter : ITypeConverter
-    {
-    }
-}
